Add some passive items + fix bug
+Thêm UI
+Thêm 4 passive mới
+ Sửa Shovel bây giờ đã có cooldown
+ Chỉnh sửa một số item

+++ Assets/Code/Enemy.cs
@@ -8,7 +8,7 @@ public class Enemy : MonoBehaviour
     public float speed;
     public float health;
     public float maxHealth;
-    public int expOnDefeat;
+    public float expOnDefeat;
     public RuntimeAnimatorController[] animController; //Dùng đề đưa animation enemies vào vd animation 1 zombie animation 2 skeleton
     public Rigidbody2D target;
     public GameObject floatingtextPrefab;
